# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi Event Alumni Pesantren API",
        "description": "API documentation for Presensi Event Alumni Pesantren system.",
        "contact": {
            "email": "admin@pesantren.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get alumni notifications",
                "description": "Returns list of notifications for the authenticated alumni user.",
                "operationId": "getNotifications",
                "responses": {
                    "200": {
                        "description": "List of notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "notifications": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Event Baru: Reuni Akbar 2025"
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "Event \"Reuni Akbar 2025\" telah dijadwalkan"
                                                            },
                                                            "body": {
                                                                "type": "string",
                                                                "example": "Event \"Reuni Akbar 2025\" telah dijadwalkan"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "upcoming_event"
                                                            },
                                                            "priority": {
                                                                "type": "string",
                                                                "example": "normal"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2026-06-01T10:00:00+00:00"
                                                            },
                                                            "is_read": {
                                                                "type": "boolean",
                                                                "example": false
                                                            },
                                                            "read_at": {
                                                                "type": "string",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "data": {
                                                                "type": "object",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get unread notification count",
                "description": "Returns the count of unread notifications for the authenticated alumni user.",
                "operationId": "getUnreadNotificationCount",
                "responses": {
                    "200": {
                        "description": "Unread notification count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification as read",
                "description": "Marks a specific notification as read for the authenticated alumni user.",
                "operationId": "markNotificationAsRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notification ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "notification": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Event Baru"
                                                        },
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Event telah dijadwalkan"
                                                        },
                                                        "body": {
                                                            "type": "string",
                                                            "example": "Event telah dijadwalkan"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "upcoming_event"
                                                        },
                                                        "priority": {
                                                            "type": "string",
                                                            "example": "normal"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2026-06-01T10:00:00+00:00"
                                                        },
                                                        "is_read": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "read_at": {
                                                            "type": "string",
                                                            "example": "2026-06-02T10:00:00+00:00"
                                                        },
                                                        "data": {
                                                            "type": "object",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "description": "Marks all unread notifications as read for the authenticated alumni user.",
                "operationId": "markAllNotificationsAsRead",
                "responses": {
                    "200": {
                        "description": "All notifications marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "updated_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message here"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}